% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/statmodlib.r
\name{confRhofaelles}
\alias{confRhofaelles}
\title{Konfidensinterval for fælles rho}
\usage{
confRhofaelles(r, n)
}
\arguments{
\item{r}{estimat for rho}

\item{n}{stikprøvestørrelse}
}
\description{
Konfidensinterval for fælles rho
}

