% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/statmodlib.r
\name{confZ}
\alias{confZ}
\title{Konfidensinterval for fishers Z}
\usage{
confZ(z, n)
}
\arguments{
\item{z}{estimat for fisherZ}

\item{n}{stikprøvestørrelse}
}
\description{
Konfidensinterval for fishers Z
}

