% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/statmodlib.r
\name{confZfaelles}
\alias{confZfaelles}
\title{Konfidensinterval for fælles Z}
\usage{
confZfaelles(z, n)
}
\arguments{
\item{z}{estimat for fishers Z}

\item{n}{stikprøvestørrelse}
}
\description{
Konfidensinterval for fælles Z
}

