% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/statmodlib.r
\name{estimates2d}
\alias{estimates2d}
\title{Estimer 2D normal fordeling}
\usage{
estimates2d(S_x, S_y, USS_x, USS_y, SP_xy, n)
}
\arguments{
\item{S_x}{summen af x'er}

\item{S_y}{summen af y'er}

\item{USS_x}{summen af x^2}

\item{USS_y}{summen af y^2}

\item{SP_xy}{summen af x*y}

\item{n}{stikprøvestørrelsen}
}
\description{
(x,y) er en observationsrække fra den 2dimensionale
normalfordeling.
}

