% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/statmodlib.r
\name{hotellingsT}
\alias{hotellingsT}
\title{Hotellings T test}
\usage{
hotellingsT(mux0, muy0, muxhat, muyhat, S, n)
}
\arguments{
\item{mux0}{mux0}

\item{muy0}{muy0}

\item{muxhat}{estimat for mu_x}

\item{muyhat}{estimat for mu_y}

\item{S}{2x2 matrix med estimat for fælles kovarians matrix (se formel (2.49))}
}
\description{
Hotellings T test for H0: (mu_x, mu_y) = (mux0, muy0)
}

