\name{statmodlib-package}
\alias{statmodlib-package}
\alias{statmodlib}
\docType{package}
\title{
Statmodlib
}
\description{
En lille R pakke for faget "Statistiske modeller 2" ved Aarhus Universitet
}
\details{
\tabular{ll}{
Package: \tab statmodlib\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-13\cr
License: \tab What license is it under?\cr
}

Se vignetter for eksempler
}
\author{
Tobias Madsen
}
\references{
Statistiske Modeller 2, Jørgen Granfeldt
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
