% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/statmodlib.r
\name{testLighedRho}
\alias{testLighedRho}
\title{Test lighed af korrelation k>=2 stikprøver}
\usage{
testLighedRho(rho, n)
}
\arguments{
\item{n}{vektor med stikprøvestørrelser}

\item{r}{vektor med estimator for rho for hver observationsrække}
}
\description{
Test lighed af korrelation k>=2 stikprøver
}

