% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/statmodlib.r
\name{testLighedZ}
\alias{testLighedZ}
\title{Test lighed af korrelation k>=2 stikprøver}
\usage{
testLighedZ(z, n)
}
\arguments{
\item{z}{vektor med estimater for fishersZ for hver observationsrække}

\item{n}{vektor med stikprøvestørrelser}
}
\description{
Test lighed af korrelation k>=2 stikprøver
}

