% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/statmodlib.r
\name{testRhoNul}
\alias{testRhoNul}
\title{Test korrelation = 0 i en stikprøve}
\usage{
testRhoNul(rho, n)
}
\arguments{
\item{rho}{estimat for rho}

\item{n}{stikprøvestørrelse}
}
\description{
Test korrelation = 0 i en stikprøve
}

